调试：文件更新消息。 修复业务中无任务标识号填写8个F无法入库的错误。 修改工具类ByteUtil，新增按照无符号整型解析数据的方法。

+++ src/main/java/cn/nci/domain/EMBLHeader.java
@@ -7,7 +7,7 @@ package cn.nci.domain;
  * @create: 2020-07-23 13:37
  */
 public class EMBLHeader {
-    private Integer taskID;
+    private Long taskID;
     private Integer dataTypeID;
     private Integer deviceID;
     private Integer date;
@@ -16,11 +16,11 @@ public class EMBLHeader {
     private Integer dataLength;
     private byte[] content;
 
-    public Integer getTaskID() {
+    public Long getTaskID() {
         return taskID;
     }
 
-    public void setTaskID(Integer taskID) {
+    public void setTaskID(Long taskID) {
         this.taskID = taskID;
     }
